+++ node_helper.js
@@ -3,6 +3,7 @@
 //
 
 const fs = require("fs")
+const { exec } = require("child_process")
 const checker = require("./components/checker.js")
 var logGA = (...args) => { /* do nothing */ }
 var NodeHelper = require("node_helper")
@@ -47,10 +48,10 @@ module.exports = NodeHelper.create({
         this.website.server()
         break
       case "ACTIVATE_ASSISTANT":
-        this.lib.activateAssistant.activate(this, payload)
+        this.activate(payload)
         break
       case "SHELLEXEC":
-        this.lib.GATools.shellExec(this, payload)
+        this.shellExec(payload)
         break
       case "GOOGLESEARCH":
         this.searchOnGoogle.search(this, payload)
@@ -116,7 +117,7 @@ module.exports = NodeHelper.create({
 
     this.searchOnGoogle = new this.lib.Assistant.GOOGLESEARCH()
 
-    await this.lib.GATools.loadRecipes(this)
+    await this.loadRecipes()
     this.sendSocketNotification("GA-INIT")
   },
 
@@ -160,9 +161,7 @@ module.exports = NodeHelper.create({
     let Libraries = []
     let GA= [
       // { "library to load" : "store library name" }
-      { "./components/assistant.js": "Assistant" },
-      { "./components/GA_Tools.js": "GATools" },
-      { "./components/activateAssistant.js": "activateAssistant" }
+      { "./components/assistant.js": "Assistant" }
     ]
    
     let website= [
@@ -232,5 +231,103 @@ module.exports = NodeHelper.create({
       this.smarthome.refreshData()
       this.smarthome.updateGraph()
     }
+  },
+
+  loadRecipes: function() {
+    if (this.config.debug) logGA = (...args) => { console.log("[GA] [RECIPES]", ...args) }
+    return new Promise(resolve => {
+      if (this.config.recipes) {
+        let replacer = (key, value) => {
+          if (typeof value == "function") {
+            return "__FUNC__" + value.toString()
+          }
+          return value
+        }
+        var recipes = this.config.recipes
+        var error = null
+        var nb_Err = 0
+        for (var i = 0; i < recipes.length; i++) {
+          try {
+            var p = require("./recipes/" + recipes[i]).recipe
+            this.sendSocketNotification("LOAD_RECIPE", JSON.stringify(p, replacer, 2))
+            console.log("[GA] [RECIPES] LOADED:", recipes[i])
+          } catch (e) {
+            error = `[FATAL] RECIPE_ERROR (${recipes[i]})`
+            console.error("[GA] [RECIPES] LOADING ERROR:", recipes[i])
+            console.error("[GA] [RECIPES] DETAIL:", e.message)
+            this.sendSocketNotification("RECIPE_ERROR", recipes[i])
+            nb_Err++
+          }
+        }
+        if (!nb_Err) console.log("[GA] Recipes loaded!")
+        else console.log(`[GA] Recipes loaded but {$nb_Err} detected!`)
+        resolve()
+      } else {
+        logGA("No Recipes to Load...")
+        resolve()
+      }
+    })
+  },
+
+  shellExec: function(payload) {
+    if (this.config.debug) logGA = (...args) => { console.log("[GA] [SHELL_EXEC]", ...args) }
+    var command = payload.command
+    if (!command) return console.error("[GA] [SHELLEXEC] no command to execute!")
+    command += (payload.options) ? (" " + payload.options) : ""
+    exec (command, (e,so,se)=> {
+      logGA("command:", command)
+      if (e) {
+        console.log("[GA] [SHELL_EXEC] Error:" + e)
+        this.sendSocketNotification("WARNING", { message: "ShellExecError"} )
+      }
+      logGA("RESULT", {
+        executed: payload,
+        result: {
+          error: e,
+          stdOut: so,
+          stdErr: se,
+        }
+      })
+    })
+  },
+
+  activate: function(payload) {
+    if (this.config.debug) logGA = (...args) => { console.log("[GA] [ACTIVATE_ASSISTANT]", ...args) }
+    logGA("QUERY:", payload)
+    var assistantConfig = Object.assign({}, this.config.assistantConfig)
+    assistantConfig.debug = this.config.debug
+    assistantConfig.lang = payload.lang
+    assistantConfig.micConfig = this.config.micConfig
+    this.assistant = new this.lib.Assistant.ASSISTANT(assistantConfig, (obj)=>{ this.sendSocketNotification("TUNNEL", obj) })
+
+    var parserConfig = {
+      responseOutputCSS: this.config.responseConfig.responseOutputCSS,
+      responseOutputURI: "tmp/responseOutput.html",
+      responseOutputZoom: this.config.responseConfig.zoom.responseOutput
+    }
+    var parser = new this.lib.Assistant.SCREENPARSER(parserConfig, this.config.debug)
+    var result = null
+    this.assistant.activate(payload, (response)=> {
+      response.lastQuery = payload
+
+      if (!(response.screen || response.audio)) {
+        if (!response.audio && !response.screen && !response.text) response.error.error = "NO_RESPONSE"
+        if (response.transcription && response.transcription.transcription && !response.transcription.done) {
+          response.error.error = "TRANSCRIPTION_FAILS"
+        }
+      }
+      if (response && response.error.audio && !response.error.message) response.error.error = "TOO_SHORT"
+
+      if (response.screen) {
+        parser.parse(response, (result)=>{
+          delete result.screen.originalContent
+          logGA("RESULT", result)
+          this.sendSocketNotification("ASSISTANT_RESULT", result)
+        })
+      } else {
+        logGA("RESULT", response)
+        this.sendSocketNotification("ASSISTANT_RESULT", response)
+      }
+    })
   }
 })
